Separated Prophero from Palaso and WordSend projects by including libraries in lib/, plus including source from one portion of WordSend, so that this project should now compile and run by itself EXCEPT for its remaining Python dependency.

+++ sepp/HTML_To_BibleReadingGuide.cs
@@ -46,7 +46,7 @@ namespace sepp
 				m_totalReadings += ReadingsForMonth(i);
 			m_readings = new Reading[m_totalReadings];
 
-			int totalLength = m_sections.Sum(info => info.Length);
+			int totalLength = (int) m_sections.Sum(info => info.Length);
 
 			int remainingLength = totalLength;
 			int iSection = 0;
